requirements.txt
appdirs==1.4.4
asgiref @ file:///tmp/build/80754af9/asgiref_1617740417847/work
attrs==21.4.0
bcrypt==3.2.0
better-exceptions==0.3.3
boto3==1.17.61
botocore==1.20.61
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
classutils==1.18.1
colorama==0.4.4
comtypes==1.1.11
configurationutil==3.7.6
conversionutil==1.4.6
cryptography==36.0.1
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1617740316746/work
django-cors-headers==3.13.0
django-countries==7.1
django-crispy-forms==1.11.2
django-embed-video==1.4.0
django-environ==0.9.0
django-heroku==0.3.1
django-mathfilters==1.0.0
django-pagination-bootstrap==2.4.4
django-ses==2.0.0
django-session-timeout==0.1.0
djangorestframework==3.13.1
dnspython==2.2.0
dominate==2.6.0
et-xmlfile==1.1.0
fdutil==1.16.0
future==0.18.2
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
jsonschema==2.5.1
logging-helper==1.9.0
logging-tree==1.9
mysqlclient==2.0.3
newsapi-python==0.2.6
openpyxl==3.0.9
paramiko==2.9.2
paystackapi==2.0.0
Pillow==8.2.0
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pycparser==2.21
PyNaCl==1.5.0
pyperclip==1.8.2
pypiwin32==223
python-dateutil==2.8.1
python-hosts==1.0.3
pyttsx3==2.90
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==303
PyYAML==6.0
requests==2.25.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
six==1.15.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
stateutil==1.4.0
tableutil==2.4.5
timingsutil==1.7.0
uiutil==1.38.0
urllib3==1.26.4
watchdog==2.1.6
whitenoise==6.2.0
wincertstore==0.2
